+++ src/flatprot/utils/style.py
